# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GoGym API",
        "description": "API REST para gerenciamento de academias, personal trainers e alunos. Permite registro de usuários (personal e aluno), autenticação via token (Sanctum), verificação de e-mail, CRUD completo de personais e alunos, e sistema de convites para vinculação entre personal e aluno.",
        "contact": {
            "name": "Gustavo",
            "email": "nao-responda@ugustavodev.com.br"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://ugustavodev.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:84/api",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/personal": {
            "get": {
                "tags": [
                    "Personais"
                ],
                "summary": "Listar todos os personais",
                "description": "Retorna uma lista paginada de todos os personal trainers cadastrados. Requer autenticação e e-mail verificado.",
                "operationId": "listarPersonais",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de personais retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PersonalResource"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginacaoLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginacaoMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "E-mail não verificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutorizado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/personal/{personal}": {
            "get": {
                "tags": [
                    "Personais"
                ],
                "summary": "Exibir um personal específico",
                "description": "Retorna os dados detalhados de um personal trainer pelo seu ID. Requer autenticação e e-mail verificado.",
                "operationId": "exibirPersonal",
                "parameters": [
                    {
                        "name": "personal",
                        "in": "path",
                        "description": "ID do personal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do personal retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PersonalResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personal não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Personais"
                ],
                "summary": "Atualizar dados de um personal",
                "description": "Atualiza parcialmente os dados de um personal trainer. Todos os campos são opcionais. Requer autenticação e e-mail verificado.",
                "operationId": "atualizarPersonal",
                "parameters": [
                    {
                        "name": "personal",
                        "in": "path",
                        "description": "ID do personal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos a serem atualizados",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telefone": {
                                        "type": "string",
                                        "example": "(11) 97777-7777",
                                        "maxLength": 20
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "feminino",
                                        "enum": [
                                            "masculino",
                                            "feminino",
                                            "nao_binario",
                                            "outro",
                                            "prefiro_nao_informar"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personal atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PersonalResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personal não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroValidacao"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Personais"
                ],
                "summary": "Inativar um personal",
                "description": "Realiza soft delete de um personal trainer. Requer autenticação e e-mail verificado.",
                "operationId": "deletarPersonal",
                "parameters": [
                    {
                        "name": "personal",
                        "in": "path",
                        "description": "ID do personal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal inativado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Personal inativado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personal não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/personal/gerar-convite": {
            "post": {
                "tags": [
                    "Convites"
                ],
                "summary": "Gerar convite para aluno",
                "description": "Gera um token de convite para vincular um aluno ao personal trainer autenticado. O aluno utilizará esse token no momento do registro. Se já existir um convite pendente para o mesmo e-mail, o existente será retornado. Requer autenticação e e-mail verificado.",
                "operationId": "gerarConvite",
                "requestBody": {
                    "description": "Dados do convite",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Silva",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Convite gerado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Convite gerado com sucesso."
                                        },
                                        "convite": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "joao@email.com"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "aB3dEfGhIjKlMnOpQrStUvWxYz1234567890abcdefghijklmnopqrstuv"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pendente",
                                                    "enum": [
                                                        "pendente",
                                                        "aceito"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroValidacao"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErroValidacao": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErroNaoAutenticado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ErroNaoAutorizado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ErroNaoEncontrado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Recurso não encontrado."
                    }
                },
                "type": "object"
            },
...
            "PersonalResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "usuario_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "telefone": {
                        "type": "string",
                        "example": "(11) 99999-9999"
                    },
                    "genero": {
                        "type": "string",
                        "example": "masculino",
                        "enum": [
                            "masculino",
                            "feminino",
                            "nao_binario",
                            "outro",
                            "prefiro_nao_informar"
                        ]
                    },
                    "cadastrado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-13 22:00:00"
                    },
                    "atualizado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-13 22:00:00"
                    },
                    "deletado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginacaoLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginacaoMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 5
                    },
                    "total": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Insira o token retornado no login/registro. Exemplo: {token}",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
}